requirements.txt
asarPy==1.0.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==39.0.2
daphne==3.0.2
defusedxml==0.7.1
Django==4.1.7
django-admin-interface==0.28.6
django-allauth==0.52.0
django-channels==0.7.0
django-colorfield==0.11.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
oauthlib==3.2.2
pexpect==4.9.0
Pillow==9.4.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pypng==0.20220715.0
python-slugify==8.0.4
python3-openid==3.2.0
qrcode==7.4.2
redis==4.5.2
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Strip==0.0.1.dev14
stripe==9.8.0
text-unidecode==1.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.15
zope.interface==6.0
